Delete LabRI_Package_shiny directory
Update to the LabRI Tool, including the option to select the confidence interval (90%, 95%, or 99%) and the correction of a bug in Fleiss’ kappa test. The issue occurred in scenarios where no results were found below the lower reference limit, causing an error in the concordance analysis of the percentage of low values (below the lower limit) between the reference interval estimated by the LabRI method and the comparative reference interval.